Move core functionality to separate module
